+++ plugins/segmentio-kafka/writer_interceptor.go
@@ -18,6 +18,8 @@
 package segmentiokafka
 
 import (
+	"context"
+
 	"github.com/segmentio/kafka-go"
 
 	"github.com/apache/skywalking-go/plugins/core/operator"
@@ -30,12 +32,18 @@ const (
 	kafkaWriterComponentID = 40
 )
 
+var internalReporterContextKey = context.Background()
+
 type WriterInterceptor struct {
 }
 
 func (w *WriterInterceptor) BeforeInvoke(invocation operator.Invocation) error {
 	writer := invocation.CallerInstance().(*kafka.Writer)
 	addr, topic := writer.Addr.String(), writer.Topic
+	ctx := invocation.Args()[0].(context.Context)
+	if internal, ok := ctx.Value(internalReporterContextKey).(bool); ok && internal {
+		return nil
+	}
 	messageList := invocation.Args()[1].([]kafka.Message)
 	operationName := kafkaWriterPrefix + topic + kafkaWriterSuffix